+++ internal/handler/handlers.go
@@ -23,13 +23,13 @@ import (
 	"github.com/0x13a/golang.cafe/internal/developer"
 	"github.com/0x13a/golang.cafe/internal/job"
 	"github.com/0x13a/golang.cafe/internal/user"
-	"github.com/0x13a/golang.cafe/pkg/database"
-	"github.com/0x13a/golang.cafe/pkg/email"
-	"github.com/0x13a/golang.cafe/pkg/imagemeta"
-	"github.com/0x13a/golang.cafe/pkg/middleware"
-	"github.com/0x13a/golang.cafe/pkg/payment"
-	"github.com/0x13a/golang.cafe/pkg/seo"
-	"github.com/0x13a/golang.cafe/pkg/server"
+	"github.com/0x13a/golang.cafe/internal/database"
+	"github.com/0x13a/golang.cafe/internal/email"
+	"github.com/0x13a/golang.cafe/internal/imagemeta"
+	"github.com/0x13a/golang.cafe/internal/middleware"
+	"github.com/0x13a/golang.cafe/internal/payment"
+	"github.com/0x13a/golang.cafe/internal/seo"
+	"github.com/0x13a/golang.cafe/internal/server"
 	"github.com/ChimeraCoder/anaconda"
 	"github.com/PuerkitoBio/goquery"
 	"github.com/bot-api/telegram"
