+++ internal/seo/seo.go
@@ -9,7 +9,7 @@ import (
 
 	"github.com/0x13a/golang.cafe/internal/company"
 	"github.com/0x13a/golang.cafe/internal/developer"
-	"github.com/0x13a/golang.cafe/pkg/database"
+	"github.com/0x13a/golang.cafe/internal/database"
 )
 
 func StaticPages() []string {
