+++ internal/server/server.go
@@ -22,12 +22,12 @@ import (
 	"github.com/0x13a/golang.cafe/internal/company"
 	"github.com/0x13a/golang.cafe/internal/developer"
 	"github.com/0x13a/golang.cafe/internal/job"
-	"github.com/0x13a/golang.cafe/pkg/config"
-	"github.com/0x13a/golang.cafe/pkg/database"
-	"github.com/0x13a/golang.cafe/pkg/email"
-	"github.com/0x13a/golang.cafe/pkg/ipgeolocation"
-	"github.com/0x13a/golang.cafe/pkg/middleware"
-	"github.com/0x13a/golang.cafe/pkg/template"
+	"github.com/0x13a/golang.cafe/internal/config"
+	"github.com/0x13a/golang.cafe/internal/database"
+	"github.com/0x13a/golang.cafe/internal/email"
+	"github.com/0x13a/golang.cafe/internal/ipgeolocation"
+	"github.com/0x13a/golang.cafe/internal/middleware"
+	"github.com/0x13a/golang.cafe/internal/template"
 	"github.com/aclements/go-moremath/stats"
 	"github.com/dustin/go-humanize"
 	"github.com/gorilla/mux"
