+++ main.go
@@ -8,13 +8,13 @@ import (
 	"github.com/0x13a/golang.cafe/internal/developer"
 	"github.com/0x13a/golang.cafe/internal/job"
 	"github.com/0x13a/golang.cafe/internal/user"
-	"github.com/0x13a/golang.cafe/pkg/config"
-	"github.com/0x13a/golang.cafe/pkg/database"
-	"github.com/0x13a/golang.cafe/pkg/email"
-	"github.com/0x13a/golang.cafe/pkg/handler"
-	"github.com/0x13a/golang.cafe/pkg/ipgeolocation"
-	"github.com/0x13a/golang.cafe/pkg/server"
-	"github.com/0x13a/golang.cafe/pkg/template"
+	"github.com/0x13a/golang.cafe/internal/config"
+	"github.com/0x13a/golang.cafe/internal/database"
+	"github.com/0x13a/golang.cafe/internal/email"
+	"github.com/0x13a/golang.cafe/internal/handler"
+	"github.com/0x13a/golang.cafe/internal/ipgeolocation"
+	"github.com/0x13a/golang.cafe/internal/server"
+	"github.com/0x13a/golang.cafe/internal/template"
 
 	"github.com/gorilla/mux"
 	"github.com/gorilla/sessions"
